Crate aws_sdk_codestarconnections

AWS CodeStar Connections

This AWS CodeStar Connections API Reference provides descriptions and usage examples of the operations and data types for the AWS CodeStar Connections API. You can use the connections API to work with connections and installations.

Connections are configurations that you use to connect AWS resources to external code repositories. Each connection is a resource that can be given to services such as CodePipeline to connect to a third-party repository such as Bitbucket. For example, you can add the connection in CodePipeline so that it triggers your pipeline when a code change is made to your third-party code repository. Each connection is named and associated with a unique ARN that is used to reference the connection.

When you create a connection, the console initiates a third-party connection handshake. Installations are the apps that are used to conduct this handshake. For example, the installation for the Bitbucket provider type is the Bitbucket app. When you create a connection, you can choose an existing installation or create one.

When you want to create a connection to an installed provider type such as GitHub Enterprise Server, you create a host for your connections.

You can work with connections by calling:

CreateConnection, which creates a uniquely named connection that can be referenced by services such as CodePipeline.

DeleteConnection, which deletes the specified connection.

GetConnection, which returns information about the connection, including the connection status.

ListConnections, which lists the connections associated with your account.

You can work with hosts by calling:

CreateHost, which creates a host that represents the infrastructure where your provider is installed.

DeleteHost, which deletes the specified host.

GetHost, which returns information about the host, including the setup status.

ListHosts, which lists the hosts associated with your account.

You can work with tags in AWS CodeStar Connections by calling the following:

ListTagsForResource, which gets information about AWS tags for a specified Amazon Resource Name (ARN) in AWS CodeStar Connections.

TagResource, which adds or updates tags for a resource in AWS CodeStar Connections.

UntagResource, which removes tags for a resource in AWS CodeStar Connections.

For information about how to use AWS CodeStar Connections, see the Developer Tools User Guide.

Re-exports

pub use config::Config;

pub use client::Client;

Modules

client

config

error

input

model

operation

output

Structs

Blob

ByteStream

Stream of binary data

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

Enums

Error

SdkError

Failed SDK Result

Statics

PKG_VERSION